Move mpidr_el1 to bitaccess macro

+++ src/bsp/rpi3/cpu.rs
@@ -1,1 +1,1 @@
-pub const BOOT_CORE_ID: usize = 0;
+pub const BOOT_CORE_ID: u64 = 0;
